requirements.txt
fastapi
uvicorn
python-dotenv
aiofiles
requests
feedparser
cron-descriptor